refactor: change folder service name
