+++ src/context/AppContext/index.tsx
@@ -1,30 +1,11 @@
-import { createContext, useState, useEffect } from "react";
+import { useState, useEffect } from "react";
 import { useAuth0 } from "@auth0/auth0-react";
 
 import linparLogo from "@assets/images/linpar.png";
 import { SectionOrientation } from "@components/layout/BoardSection/types";
 
-import {
-  IAppContext,
-  AppContextProviderProps,
-  IClient,
-  IPreferences,
-} from "./types";
-
-export const AppContext = createContext<IAppContext>({
-  user: {
-    username: "",
-    id: "",
-    company: "",
-    operator: { name: "", logo: "" },
-    preferences: {
-      boardOrientation: "vertical",
-      showPinnedOnly: false,
-    },
-  },
-  handleClientChange: () => {},
-  updatePreferences: () => {},
-});
+import { AppContext } from "./AppContext";
+import { AppContextProviderProps, IClient, IPreferences } from "./types";
 
 export default function AppContextProvider(props: AppContextProviderProps) {
   const { children } = props;
@@ -41,7 +22,7 @@ export default function AppContextProvider(props: AppContextProviderProps) {
   });
 
   const initialClientLogo = localStorage.getItem("clientLogo") || linparLogo;
-  const[clientLogo, setClientLogo] = useState<string>(initialClientLogo);
+  const [clientLogo, setClientLogo] = useState<string>(initialClientLogo);
 
   function handleClientChange(client: IClient) {
     const { sigla, logo } = client;
@@ -68,7 +49,7 @@ export default function AppContextProvider(props: AppContextProviderProps) {
 
   const company = clientSigla;
 
-  const userContext: IAppContext = {
+  const userContext = {
     user: {
       username: `${user?.name}`,
       id: "abc123",
